(SMAR-6420) Allow production and development branches to be specified
Before this commit production and development branches couldn't be
changed from the defaults. Some repos do not use `develop` as their
development branch and instead of forcing other repos to change the
branching strategy to handle their code this commit makes it so that we
can adapt to whatever each repo needs.