(maint) Fix decoding errors in Python 3.x
Before this commit anything that was being returned via a popen result
was a bytestring which is problematic because string some string methods
do not work on bytestrings.

This commit makes sure that methods which return stdout are decoded to
strings instead of bytestrings so that they may be properly worked upon.